move simulator to simulator.py
